requirements.txt
kafka-python==2.0.2
pandas==2.0.3
pyspark==3.5.0
scikit-learn==1.3.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
joblib==1.3.2
openpyxl==3.1.2
pydantic==2.5.0
numpy==1.24.3
py4j==0.10.9.7